Move rest of view model to same file
